metaphor/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
certifi==2023.7.22
charset-normalizer==3.3.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
etelemetry==0.3.0
filelock==3.12.4
fitz==0.0.1.dev2
frozenlist==1.4.0
future==0.18.3
httplib2==0.22.0
idna==3.4
isodate==0.6.1
joblib==1.3.2
looseversion==1.3.0
lxml==4.9.3
metaphor-python==0.1.16
multidict==6.0.4
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
pathlib==1.0.1
pdfreader==0.1.12
Pillow==10.0.1
prov==2.0.0
pycryptodome==3.19.0
pydot==1.4.2
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2
